pyproject.toml
[tool.poetry]
name = "talk_dev_to_me_backend"
version = "0.1.0"
description = ""
authors = ["NickDorkins <nick.dorkins@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.1"
djangorestframework = "^3.12.4"
docker = "^5.0.0"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
djangorestframework-simplejwt = "^4.6.0"
gunicorn = "^20.1.0"
whitenoise = "^5.2.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
